Update revision to apache
The previous version of apacha led to dep failures
when installing mod_wsgi.

This patch updates the revision of the upstream
apache module to include a patch that I added
that resolves this dep issue (where .conf files
for a module were added before the package
was installed)

This also requires that we update the version of
concat to 1.1.1 to get a fixed bug.