Udate the way that a mon config is determined
This patch updates how a mon_config is determine to avoid
the need to disovery both short hostname as well as ip address
for cep mon nodes.

It devides the address info into leader vs. members which are
each an address per node.

It then checks the hostname and address for public interface
against those ips and replaces the mon name with the hostname
for the one that matches the localhost.