Add puppet master
This commit adds a puppet master to
the etcd bootstrap node that will
only serve as a CA server.

the purpose of the CA server is to
have the cluster generate it's own set
of certs that can be used to authenticate
hosts in consul.

This commits adds both the puppet and consul
specific configuration.