add magic ip to openstack object code
currently, in the at env, the fact that identity.jiocloud.com
already resolves to the undercloud address is causing the ocdb
role install to take forever. This commit adds the default
resovled address for this address to a special list so that
it is ignored in terms of determining if the address is
reachable.