bind puppet/validation checks to services
This patch creates services for puppet/validation
checks and binds the existing ttl checks to them.

This is done so that the state of those patches does
not interfere with other services being run. In particular,
it seems that with consul 5, the default state of service
checks is critical until they recieve an update. That means
that patch is required for builds to work with consul 5.

The reason that this patch creates services separately from
checks is so that the checks can retain the same names so
as to not break jiocloud-python (checks embedded in services
have a different name, even if the id/name for those checks
are set explicitly)