Code to create fip pool using different methods
FIP pools can be of two types.

1. Shared to all tenants - this is setup using neutron apis, and it is shared to
all tenants.
2. Shared to specific tenants only - this need to be setup using contrail api,
and it can be shared only to specific tenants. so this can be a private floating
IPs to one or more tenants.

Created one defined resource to create both the type of fip pools. So a hash of
fip pools to be provided using hiera to the class rjil::neutron::contrail and it
call the define rjil::neutron::contrail::fip_pool to create the fips using
appropriate apis.

Changed Puppetfile to point to harish's repo for contrail to test.