up node build time to 4500 seconds
After the move to volumes for disk in the cloud
environment, it takes ceph 30-40 minutes to gets
it's pages in a clean state, this causes all builds
to fail.

This patch ups the build time allowed for the environment
to 1 hour 15 minutes to ensure this does not cause all builds
to fail.