Create seed_orchestrator definition
This commit abstracts the seed based deployment
pattern into a reusable defined resource type.

The hope is that we can abstract our orchestration
routines in a way that makes implementation for
all service that require a single seed node for
bootstrapping clusters can easily be reimplemted
for all future usages.