Scale zookeeper
Add code that can build out a zookeeper cluster.
Scaling zookeeper is crazy hard, this patch does
the following:

- blocks until we have min_numbers of zookeeper
instances which are tracked via the pre-zookeeper
service (which is used to indicate that a machine
will be a zookeeper instance)
- adds a zookeeper service, adds servers param to
zookeeper. Adds functions that can be used to create
the server list.
- customizes datastore so that it will wipe out old
data to avoid clustering issues.
- non-seed nodes let the seed be created first to
avoid split brain on bootstrapping