Add the concept of an override_repo
This allows users to specify an external location where an
apt repo can be uploaded (like through swift, or on an apache server).

This repo can be downloaded and tested as a part of a normal build.

NOTE: this patch uses a module called archive which depends on
some packages, since it is setting up an apt::source, it
cannot install those packages first (b/c apt::source comes
before all packages). For this reason, those deps are installed
from userdata which might be less than ideal.