Pre-download glance image
In kilo, /usr/bin/glance does not support
remote images for download. This patch ensures
that our desired glance images are copied locally
before being added to the glance db.

This also resolves the issue of the glance images
not going through the proxy and probably also means
that we can remove the fip from the ocdb role.