Add service selection code for os objects
This code adds selection code that can be used
to determine what objects to create based on
what services exist on a given node.

This code is intended to make it easier
to support just installing subsets of the
openstack components for development.

Also ensure that tenant networks respect enable_neutron