(DO NOT MERGE) Add auth_url env var
This is a tmp commit and not meant to be
the real solution.

basically, the new keystone types use keystone
v3, but will also use the endpoint you specify
in /root/openrc (it treats openrc as it's config
file)

In our case, it causes an issue b/c we have v2
endppoints there b/c we use v2 for validation tests.

We probably need to completely move from v2 to v3
for everything or at least use different config files
for our own validation checks.