Support pausing puppet
Presently the only way to disable puppet is to disable the cron
This provides for a way to check in config if puppet has to be run or not
Right now the python-jiocloud provider only has enabled_puppet True or False
Going forward, we will add the ability to "pause" puppet till a given time as well.
Though this remaining pausing change will only be python-jiocloud and nothing in puppet-rjil.

This feature is intended to support setting phased rollouts through keys in consul
to support rolling upgrades.