Do not run valiation tests before Puppet runs
This previous commit ran valiation tests before
Puppet to ensure they would always run even if Puppet
was disabled.

The problem with this is that validation checks exit
if validation fails. This meant that it guarenteed that
machines could not bootstrap in cases where validation checks
were not passing b/c validation failed, never allowing Puppet
to run.

This commit updates the code, by not effecting the current order
of things and just ensuring that valiation checks are run in the
conditional statement that exists with return code 9.