Fail if check_puppet retuns 1
It was possbile for check_puppet to receive
false positives in cases where the program exited
with 1 (indicated a python error like an uncaught
exception).

This commit ensures that returns of 1 are handled and
that Puppet is not run in those cases since we cannot assume
that Puppet is supposed to run if our communication with
consul fails.