Delete ports form leftover networks
Add a force delete method that can actually
delete leftover networks. This method explicitly
delete leftoever networks and any ports that might
exist for those networks.

The previous code did not work for the case where
a network/subnet/port had been created and not
cleaned up.

In this case, the network cannot be deleted b/c
it has a port assigned.