add layout to hiera
Layout is the concept that is used
to model a deployment topology.

In order to be able to properly
support multiple layouts for openstack,
we need a place to put layout specific
configuration overrides.

This patch adds a fact for layout via
facter.d and adds an additional hierarchy
that uses that fact to allow layout
specfic hiera overrides.

This patch is a part of the work to create
a layout that just deploys keystone.